🐳 chore: Move VisualQueryEditor into independent folder
