+++ src/components/QueryEditor/VisualQueryEditor/AddSegment.tsx
