+++ src/components/QueryEditor/VisualQueryEditor/NonStateSegmentInput.tsx
