+++ src/components/QueryEditor/VisualQueryEditor/RemoveSegment.tsx
