+++ src/components/QueryEditor/VisualQueryEditor/index.tsx
@@ -1,0 +1,3 @@
+import { VisualQueryEditor } from './VisualQueryEditor';
+
+export default VisualQueryEditor;
